Split Dao classes.
