+++ src/main/java/bio/terra/dao/DatasetTableDao.java
@@ -12,8 +12,8 @@ import org.springframework.stereotype.Repository;
 
 import javax.sql.DataSource;
 import java.sql.Connection;
-import java.sql.ResultSet;
 import java.sql.SQLException;
+import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
@@ -23,67 +23,104 @@ import java.util.function.Function;
 import java.util.stream.Collectors;
 
 @Repository
-public class DatasetTableDao extends TableDaoBase {
+public class DatasetTableDao {
 
     private static final Logger logger = LoggerFactory.getLogger(DatasetTableDao.class);
 
     private static final String sqlInsertTable = "INSERT INTO dataset_table " +
         "(name, dataset_id, primary_key) VALUES (:name, :dataset_id, :primary_key)";
+    private static final String sqlInsertColumn = "INSERT INTO dataset_column " +
+        "(table_id, name, type, array_of) VALUES (:table_id, :name, :type, :array_of)";
+    private static final String sqlSelectTable = "SELECT id, name, primary_key FROM dataset_table " +
+        "WHERE dataset_id = :dataset_id";
+    private static final String sqlSelectColumn = "SELECT id, name, type, array_of FROM dataset_column " +
+        "WHERE table_id = :table_id";
 
     private final DataSource jdbcDataSource;
     private final NamedParameterJdbcTemplate jdbcTemplate;
 
-
     @Autowired
     public DatasetTableDao(DataRepoJdbcConfiguration jdbcConfiguration, NamedParameterJdbcTemplate jdbcTemplate) {
-        super(jdbcTemplate, "dataset_table", "dataset_column", "dataset_id");
         this.jdbcDataSource = jdbcConfiguration.getDataSource();
         this.jdbcTemplate = jdbcTemplate;
     }
 
-    @Override
-    protected UUID createTable(UUID parentId, Table table) {
+    // Assumes transaction propagation from parent's create
+    public void createTables(UUID parentId, List<Table> tableList) {
         MapSqlParameterSource params = new MapSqlParameterSource();
         DaoKeyHolder keyHolder = new DaoKeyHolder();
-
-        params.addValue("name", table.getName());
         params.addValue("dataset_id", parentId);
 
-        List<String> naturalKeyStringList = table.getPrimaryKey()
-            .orElseThrow(() -> new IllegalStateException("Dataset table " + table.getName() + " has no primary key"))
-            .stream()
-            .map(Column::getName)
-            .collect(Collectors.toList());
-        try (Connection connection = jdbcDataSource.getConnection()) {
-            params.addValue("primary_key", DaoUtils.createSqlStringArray(connection, naturalKeyStringList));
-        } catch (SQLException e) {
-            logger.error("Failed to convert primary key list to SQL array", e);
-            throw new IllegalArgumentException("Failed to convert primary key list to SQL array", e);
-        }
+        for (Table table : tableList) {
+            params.addValue("name", table.getName());
 
-        jdbcTemplate.update(sqlInsertTable, params, keyHolder);
-        return keyHolder.getId();
+            List<String> naturalKeyStringList = table.getPrimaryKey()
+                .orElse(Collections.emptyList())
+                .stream()
+                .map(Column::getName)
+                .collect(Collectors.toList());
+            try (Connection connection = jdbcDataSource.getConnection()) {
+                params.addValue("primary_key", DaoUtils.createSqlStringArray(connection, naturalKeyStringList));
+            } catch (SQLException e) {
+                logger.error("Failed to convert primary key list to SQL array", e);
+                throw new IllegalArgumentException("Failed to convert primary key list to SQL array", e);
+            }
+
+            jdbcTemplate.update(sqlInsertTable, params, keyHolder);
+
+            UUID tableId = keyHolder.getId();
+            table.id(tableId);
+            createColumns(tableId, table.getColumns());
+        }
     }
 
-    @Override
-    protected Table retrieveTable(ResultSet rs) throws SQLException {
-        Table table = new Table()
-            .id(rs.getObject("id", UUID.class))
-            .name(rs.getString("name"));
+    private void createColumns(UUID tableId, Collection<Column> columns) {
+        MapSqlParameterSource params = new MapSqlParameterSource();
+        params.addValue("table_id", tableId);
+        DaoKeyHolder keyHolder = new DaoKeyHolder();
+        for (Column column : columns) {
+            params.addValue("name", column.getName());
+            params.addValue("type", column.getType());
+            params.addValue("array_of", column.isArrayOf());
+            jdbcTemplate.update(sqlInsertColumn, params, keyHolder);
+            UUID columnId = keyHolder.getId();
+            column.id(columnId);
+        }
+    }
 
-        List<String> primaryKey = DaoUtils.getStringList(rs, "primary_key");
-        List<Column> columns = retrieveColumns(table);
-        Map<String, Column> columnMap = columns
-            .stream()
-            .collect(Collectors.toMap(Column::getName, Function.identity()));
+    public List<Table> retrieveTables(UUID parentId) {
+        MapSqlParameterSource params = new MapSqlParameterSource().addValue("dataset_id", parentId);
+        return jdbcTemplate.query(sqlSelectTable, params, (rs, rowNum) -> {
+            Table table = new Table()
+                .id(rs.getObject("id", UUID.class))
+                .name(rs.getString("name"));
 
-        List<Column> naturalKeyColumns = Optional.ofNullable(primaryKey)
-            .orElse(Collections.emptyList())
-            .stream()
-            .map(columnMap::get)
-            .collect(Collectors.toList());
-        table.primaryKey(naturalKeyColumns);
+            List<String> primaryKey = DaoUtils.getStringList(rs, "primary_key");
+            List<Column> columns = retrieveColumns(table);
+            Map<String, Column> columnMap = columns
+                .stream()
+                .collect(Collectors.toMap(Column::getName, Function.identity()));
 
-        return table.columns(columns);
+            List<Column> naturalKeyColumns = Optional.ofNullable(primaryKey)
+                .orElse(Collections.emptyList())
+                .stream()
+                .map(columnMap::get)
+                .collect(Collectors.toList());
+            table.primaryKey(naturalKeyColumns);
+
+            return table.columns(columns);
+        });
+    }
+
+    private List<Column> retrieveColumns(Table table) {
+        return jdbcTemplate.query(
+            sqlSelectColumn,
+            new MapSqlParameterSource().addValue("table_id", table.getId()), (rs, rowNum) ->
+                new Column()
+                    .id(rs.getObject("id", UUID.class))
+                    .table(table)
+                    .name(rs.getString("name"))
+                    .type(rs.getString("type"))
+                    .arrayOf(rs.getBoolean("array_of")));
     }
 }
